feat(tarea_5): finish the course

+++ tarea_5/Rolling_Hashing.cpp
@@ -1,8 +1,11 @@
-#include<bits/stdc++.h>
+#include <iostream>
+#include <vector>
+
+
 using namespace std;
 const int MAXN = 1e6;
-const ll z = 131, MOD = 1e9+21;
-ll Z[MAXN+1];
+const long long z = 131, MOD = 1e9+21;
+long long Z[MAXN+1];
 
 //precalcula las potencias de (z^i)%MOD (importante hacer antes de hacer rolling hashing)
 void fillZ() {
@@ -12,7 +15,7 @@ void fillZ() {
 }
 
 struct RollingHashing {
-  vector<ll> H;
+  vector<long long> H;
 
   //crea los hash del string s
   RollingHashing(string &S) {
@@ -23,9 +26,9 @@ struct RollingHashing {
       H[i] = (S[i] + z*H[i+1])%MOD;
   }
 
-  //entrega el hash del string s desde el indice i de tamaño k
-  ll hash(int i, int k) {
-    ll ans = (H[i] - Z[k]*H[i+k]%MOD);
+  //entrega el hash del string s desde el indice i de tamaï¿½o k
+  long long hash(int i, int k) {
+    long long ans = (H[i] - Z[k]*H[i+k]%MOD);
     return ans >= 0 ? ans : ans + MOD;
   }
 };
